Provide higher-level API for running task
